third_party/WebKit/LayoutTests/inspector-protocol/css/css-create-stylesheet-and-add-rule.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="css-protocol-test.js"></script>
<link rel="stylesheet" type="text/css" href="resources/stylesheet.css"></link>
<script type="text/javascript">
function test()
{
    var frameId;
    var styleSheetHeader;
    var documentNodeId;

    InspectorTest.requestDocumentNodeId(onDocumentNodeId);

    function onDocumentNodeId(nodeId)
    {
        documentNodeId = nodeId;
        InspectorTest.requestMainFrameId(step2);
    }

    function step2(mainFrameId)
    {
        frameId = mainFrameId;
        InspectorTest.eventHandler["CSS.styleSheetAdded"] = styleSheetAdded;
        InspectorTest.sendCommandOrDie("CSS.enable", {}, function() {})
    }

    function styleSheetAdded(response)
    {
        var header = response.params.header;
        var urlString = header.sourceURL ? " (" + InspectorTest.displayName(header.sourceURL) + ")" : "";
        InspectorTest.log("Style sheet added: " + header.origin + urlString);
        if (styleSheetHeader)
            return;

        styleSheetHeader = header;
        InspectorTest.loadAndDumpMatchingRules(documentNodeId, "#inspected", step3);
    }

    function step3()
    {
        InspectorTest.log("Adding a rule to the existing stylesheet.");
        InspectorTest.addRule(styleSheetHeader.styleSheetId, false, {
            location: { startLine: 0, startColumn: 0, endLine: 0, endColumn: 0 },
            ruleText: "#inspected {}",
        }, step4);
    }

    function step4()
    {
        InspectorTest.loadAndDumpMatchingRules(documentNodeId, "#inspected", step5);
    }

    function step5()
    {
        InspectorTest.log("Creating inspector stylesheet.");
        InspectorTest.sendCommandOrDie("CSS.createStyleSheet", { frameId: frameId }, step6);
    }

    function step6(result)
    {
        InspectorTest.log("Adding a rule to the inspector stylesheet.");
        InspectorTest.addRule(result.styleSheetId, false, {
            location: { startLine: 0, startColumn: 0, endLine: 0, endColumn: 0 },
            ruleText: "#inspected {}",
        }, step7);
    }

    function step7()
    {
        InspectorTest.loadAndDumpMatchingRules(documentNodeId, "#inspected", step8);
    }

    function step8()
    {
        InspectorTest.completeTest();
    }
};

window.addEventListener("DOMContentLoaded", function () {
    runTest();
}, false);

</script>
</head>
<body>
<div id="inspected">Inspected contents</div>
</body>
</html>
