third_party/WebKit/LayoutTests/inspector-protocol/css/css-create-stylesheet-and-add-rule-expected.txt

Inspected contents
Style sheet added: regular (stylesheet.css)
Dumping matched rules: 
*#inspected* {    regular
    color: red;
}
Adding a rule to the existing stylesheet.
==== Style sheet text ====
#inspected {}#inspected {
    color: red;
}
Dumping matched rules: 
*#inspected* {    regular
}
*#inspected* {    regular
    color: red;
}
Creating inspector stylesheet.
Style sheet added: inspector
Adding a rule to the inspector stylesheet.
==== Style sheet text ====
#inspected {}
Dumping matched rules: 
*#inspected* {    regular
}
*#inspected* {    regular
    color: red;
}
*#inspected* {    inspector
}
